feat: Generated unit tests

+++ gen/run_generation.py
@@ -1,15 +1,16 @@
 import os
-import requests
-from gen.generator import generate_clients
-from gen.helpers import generate_models, ruff_format
 import shutil
 
+import requests
+
+from gen.client_generator import generate_clients
+from gen.helpers import generate_models, put_inits, ruff_format
 
 WORKDIR = os.getcwd()
 
 SERVICES = {
     "Entitlements": "https://community.opengroup.org/osdu/platform/security-and-compliance/entitlements/-/raw/master/docs/api/entitlements_openapi.yaml?ref_type=heads",
-    "SMDS": "https://community.opengroup.org/osdu/platform/domain-data-mgmt-services/seismic/seismic-dms-suite/seismic-store-service/-/raw/master/app/sdms/docs/api/openapi.osdu.yaml?ref_type=heads",
+    "SDMS": "https://community.opengroup.org/osdu/platform/domain-data-mgmt-services/seismic/seismic-dms-suite/seismic-store-service/-/raw/master/app/sdms/docs/api/openapi.osdu.yaml?ref_type=heads",
     "Storage": "https://community.opengroup.org/osdu/platform/system/storage/-/raw/master/docs/api/storage_openapi.yaml?ref_type=heads",
     "Search": "https://community.opengroup.org/osdu/platform/system/search-service/-/raw/master/docs/api/search_openapi.yaml?ref_type=heads",
     "Indexer": "https://community.opengroup.org/osdu/platform/system/indexer-service/-/raw/master/docs/api/indexer_openapi.yaml?ref_type=heads",
@@ -22,26 +23,21 @@ SERVICES = {
     "Dataset": "https://community.opengroup.org/osdu/platform/system/dataset/-/raw/master/docs/dataset.swagger.yaml?ref_type=heads",
     "Legal": "https://community.opengroup.org/osdu/platform/security-and-compliance/legal/-/raw/master/docs/api/legal_openapi.yaml?ref_type=heads",
     "Policy": "https://community.opengroup.org/osdu/platform/security-and-compliance/policy/-/raw/master/docs/openapi.yaml?ref_type=heads",
-    "Secret": "https://community.opengroup.org/osdu/platform/security-and-compliance/secret/-/raw/main/docs/api/secret_openapi.json?ref_type=heads"
+    "Secret": "https://community.opengroup.org/osdu/platform/security-and-compliance/secret/-/raw/main/docs/api/secret_openapi.json?ref_type=heads",
+    "Wellbore": "https://community.opengroup.org/osdu/platform/domain-data-mgmt-services/wellbore/wellbore-domain-services/-/raw/master/spec/generated/openapi.json?ref_type=heads",
+    "WellDelivery": "https://community.opengroup.org/osdu/platform/domain-data-mgmt-services/well-delivery/well-delivery/-/raw/master/docs/api/swagger.yaml?ref_type=heads",
+    "RAFS": "https://community.opengroup.org/osdu/platform/domain-data-mgmt-services/rock-and-fluid-sample/rafs-ddms-services/-/raw/main/docs/spec/openapi.json?ref_type=heads"
 }
-
-
-def put_inits(path: str):
-    _path = path.replace(WORKDIR, "")[1:].split("/")
-    while len(_path) > 1:
-        open(
-            os.path.join(*_path, "__init__.py"), "w"
-        ).close()
-        _path.pop()
 
 
 def remove_path_if_empty(path: str):
     _path = os.path.join(path, "__init__.py")
-    with open(_path, 'r') as fp:
-        lines = len(fp.readlines())
+    if os.path.exists(_path):
+        with open(_path, 'r') as fp:
+            lines = len(fp.readlines())
 
-    if lines == 0:
-        shutil.rmtree(path)
+        if lines == 0:
+            shutil.rmtree(path)
 
 
 if __name__ == "__main__":
@@ -50,9 +46,9 @@ if __name__ == "__main__":
         url = SERVICES[name]
         module_path = os.path.join(WORKDIR, "osdu_client", "services", name.lower())
         models_path = os.path.join(module_path, "models")
-
-        os.makedirs(module_path, exist_ok=True)
-        put_inits(module_path)
+        rel_module_path = module_path.replace(WORKDIR, ".")
+        os.makedirs(rel_module_path, exist_ok=True)
+        put_inits(rel_module_path)
 
         swagger_path = os.path.join(module_path, "swagger.yaml")
 
@@ -65,17 +61,18 @@ if __name__ == "__main__":
 
                 f.write(response.text)
 
-        generate_models(
-            swagger_path,
-            models_path+".py",
-        )
+        if not (os.path.exists(models_path) or os.path.exists(models_path+".py")):
+            generate_models(
+                swagger_path,
+                models_path+".py",
+            )
         generate_clients(
             swagger_path,
             name,
-            dump_file=os.path.join(WORKDIR, "osdu_client", "services", name.lower(), "client.py")
+            dump_file_path=os.path.join(WORKDIR, "osdu_client", "services", name.lower(), "client.py")
         )
         ruff_format(
-            os.path.join(WORKDIR, "osdu_client", "services", name.lower(), "client.py")
+            os.path.join(WORKDIR, "osdu_client", "services", name.lower())
         )
         # os.remove(swagger_path)
-        # remove_path_if_empty(models_path)
+        remove_path_if_empty(models_path)
